requirements.txt
altair==4.2.2
altgraph==0.17.2
ansicon==1.89.0
anyio==3.6.2
appdirs==1.4.4
asgiref==3.5.2
attrs==22.2.0
bidict==0.22.1
blessed==1.19.1
blinker==1.5
bottle==0.12.23
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
dataclasses-json==0.5.7
decorator==5.1.1
Django==4.1.2
docutils==0.17.1
entrypoints==0.4
fastapi==0.88.0
fastapi-socketio==0.0.9
Flask==2.2.2
flyingcircus==0.1.4.0
fonttools==4.38.0
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
markdown-it-py==2.2.0
markdown2==2.4.7
MarkupSafe==2.1.2
marshmallow==3.17.0
marshmallow-enum==1.5.1
matplotlib==3.6.3
mdurl==0.1.2
mypy-extensions==0.4.3
nicegui==1.1.3
numpy==1.23.3
packaging==21.3
pandas==1.5.0
pefile==2022.5.30
Pillow==9.4.0
protobuf==3.20.3
pscript==0.7.7
psycopg2 @ file:///D:/Download/%E2%98%85%E4%BC%9A%E7%A4%BE/psycopg2-2.9.3-cp310-cp310-win_amd64.whl
pyarrow==11.0.0
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.2
pytk==0.0.2.1
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.2
rich==13.3.2
semver==2.13.0
setuptools-scm==7.0.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.22.0
streamlit==1.20.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typing-inspect==0.7.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.20.0
validators==0.20.0
vbuild==0.8.1
watchdog==3.0.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
Werkzeug==2.2.2
zipp==3.15.0
